# flask/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Access API",
        "description": "API endpoints for UGV-Based Disaster Management System",
        "version": "1.0"
    },
    "basePath": "/",
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Bearer token"
        }
    },
    "paths": {
...
        "/api/missions": {
            "post": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission Creation [ADMIN]",
                "description": "Create a new mission with the provided details",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Mission data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Mission ABC",
                                    "required": true
                                },
                                "device_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f268f122c3647c263a1a79"
                                    }
                                },
                                "user_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f64a53767178df500184ec"
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mission created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Mission created successfully."
                                },
                                "mission_id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - Device/User not found"
                    },
                    "409": {
                        "description": "Conflict - Mission name is already taken"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/missions/{mission_id}": {
            "get": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission Information Retrievement",
                "description": "Retrieve information about a specific mission",
                "parameters": [
                    {
                        "name": "mission_id",
                        "in": "path",
                        "description": "ID of the mission to retrieve",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission information retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Mission ABC"
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-03-14 06:06:28.018762"
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-03-16 06:06:28.018762"
                                },
                                "status": {
                                    "type": "int",
                                    "example": 1
                                },
                                "devices": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "6055a0ae80e90e08641e22ef"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Device1"
                                            },
                                            "type": {
                                                "type": "int",
                                                "example": 1
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - Mission not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission Info Update [ADMIN]",
                "description": "update name, device_ids, user_ids, start_date or end_date for the Mission",
                "parameters": [
                    {
                        "name": "mission_id",
                        "in": "path",
                        "description": "ID of the mission to update",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Mission data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Mission ABC",
                                    "required": true
                                },
                                "device_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f268f122c3647c263a1a79"
                                    }
                                },
                                "user_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f64a53767178df500184ec"
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission information is updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Mission information is updated successfully."
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Mission ABC"
                                },
                                "device_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f268f122c3647c263a1a79"
                                    }
                                },
                                "user_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "65f64a53767178df500184ec"
                                    }
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2020-02-01T10:11:30"
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2020-02-04T06:00:34"
                                }
                            }
                        },
                        "400": {
                            "description": "Bad Request"
                        },
                        "401": {
                            "description": "Unauthorized Error - Admin access is required"
                        },
                        "403": {
                            "description": "Forbidden Error"
                        },
                        "404": {
                            "description": "Not Found - Mission not found"
                        },
                        "409": {
                            "description": "Conflict"
                        },
                        "500": {
                            "description": "Internal Server Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/missions/all": {
            "get": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission List Retrievement [ADMIN]",
                "description": "Get list of missions with pagenation for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page-number",
                        "type": "integer",
                        "required": false,
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "in": "query",
                        "name": "page-size",
                        "type": "integer",
                        "required": false,
                        "description": "Number of items per page",
                        "default": 5
                    },
                    {
                        "in": "query",
                        "name": "name",
                        "type": "string",
                        "required": false,
                        "description": "filter for mission name",
                        "default": "mission"
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "type": "integer",
                        "required": false,
                        "description": "filter for mission status",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "6055a0ae80e90e08641e22ef"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "mission A"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/missions/count": {
            "get": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission Count [ADMIN]",
                "description": "Get the count of mission with filters for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "description": "filter for the user status",
                        "type": "array",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "array",
                                    "example": [
                                        1,
                                        3
                                    ]
                                },
                                "count": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/missions/{mission_id}/{command}": {
            "put": {
                "tags": [
                    "Mission Management"
                ],
                "summary": "Mission Status Update [ADMIN]",
                "description": "update status for the Mission\n\nCommands:\n-start\n-pause\n-continue\n-cancel\n-end",
                "parameters": [
                    {
                        "name": "mission_id",
                        "in": "path",
                        "description": "ID of the mission to update",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    },
                    {
                        "name": "command",
                        "in": "path",
                        "description": "command for the mission to update the status",
                        "required": true,
                        "type": "string",
                        "example": "start"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission information is updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Mission status is updated successfully."
                                }
                            }
                        },
                        "400": {
                            "description": "Bad Request"
                        },
                        "401": {
                            "description": "Unauthorized Error - Admin access is required"
                        },
                        "403": {
                            "description": "Forbidden Error"
                        },
                        "404": {
                            "description": "Not Found - Mission not found"
                        },
                        "409": {
                            "description": "Conflict"
                        },
                        "500": {
                            "description": "Internal Server Error"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    }
}